refactor: AppView와 extension
